Compare performance of appending new data vs rewriting
The above are different approaches to building the DataFrames which represent the data received as we move forward in time.

The append method adds new rows to the DataFrames as we move forward in time, while the rewrite method initalises a DataFrame with the indexes (timestamps) of the data which will be received.

It appears that the append method is significantly faster than the rewrite method. I'm unsure why this is the case as I understand appending would require an additional step of memory allocation.