Refactor DataHandler and EventQueue
Tidied up the tests at the end of the file.

Added a generator for future data which I realised is useless now because
loading the queue isn't the bottleneck. However, this should be useful
when we are dealing with many sources of data over longer periods of time.

Removed the EventQueue method for loading future data as a list
comprehension in main is faster.